Symlinks taskdata into taskpublisher
