+++ taskpublisher/src/main/kotlin/io/taskdata
@@ -1,0 +1,1 @@
+../../../../../shared_packages/taskdata
