--- taskpublisher/src/main/kotlin/io/taskdata/Task.kt
@@ -1,16 +1,0 @@
-package io.taskdata
-
-import kotlinx.serialization.Serializable
-
-enum class Status {
-    COMPLETED, RUNNING, FAILED, SCHEDULED
-}
-
-@Serializable
-data class Task(
-    val name: String,
-    val fileName: String,
-    val minute: Int?,
-    val hour: Int?,
-    val status: Status = Status.SCHEDULED
-)
