--- taskpublisher/src/main/kotlin/io/taskdata/TaskRepository.kt
@@ -1,6 +1,0 @@
-package io.taskdata
-
-interface TaskRepository {
-    suspend fun allTasks(): List<Task>
-    suspend fun queryTasks(hour: Int, minute: Int): List<Task>
-}
